refactor: split `GetTotalRAM()` into `common/system_ram.{cpp,h}`
Move the OS-dependent total RAM query out of `common/system.cpp` into a dedicated `common/system_ram.cpp` translation unit and declare it in `common/system_ram.h`.
